+++ Beat iOS/Beat-iOS-Bridging-Header.h
@@ -1,0 +1,23 @@
+//
+//  Beat-iOS-Bridging-Header.h
+//  Beat iOS
+//
+//  Created by Lauri-Matti Parppei on 14.5.2022.
+//  Copyright © 2022 Lauri-Matti Parppei. All rights reserved.
+//
+
+#ifndef Beat_iOS_Bridging_Header_h
+#define Beat_iOS_Bridging_Header_h
+
+#import "BeatDocumentSettings.h"
+#import "Line.h"
+#import "OutlineScene.h"
+//#import "BeatAppDelegate.h"
+#import "BeatEditorDelegate.h"
+//#import "BeatNotepad.h"
+//#import "BeatLocalization.h"
+// #import "ThemeManager.h"
+// #import "BeatPrintDialog.h"
+#import "BeatPaginator.h"
+
+#endif /* Beat_iOS_Bridging_Header_h */