+++ Beat iOS/iOSDocument.swift
@@ -1,0 +1,22 @@
+//
+//  Document.swift
+//  Beat iOS
+//
+//  Created by Lauri-Matti Parppei on 14.5.2022.
+//  Copyright © 2022 Lauri-Matti Parppei. All rights reserved.
+//
+
+import UIKit
+
+class iOSDocument: UIDocument {
+    
+    override func contents(forType typeName: String) throws -> Any {
+        // Encode your document with an instance of NSData or NSFileWrapper
+        return Data()
+    }
+    
+    override func load(fromContents contents: Any, ofType typeName: String?) throws {
+        // Load your document from contents
+    }
+}
+