+++ Beat/BeatAppDelegate.m
@@ -91,7 +91,7 @@
 
 - (void) awakeFromNib {
 	// [NSUserDefaults.standardUserDefaults setBool:NO forKey:@"NSConstraintBasedLayoutVisualizeMutuallyExclusiveConstraints"];
-		
+	
 #ifdef ADHOC
 	// Ad hoc vector
 	NSLog(@"# ADHOC");