+++ Beat/User Interface/Editor View/BeatTextView.h
@@ -78,8 +78,6 @@ typedef NS_ENUM(NSInteger, BeatEditorMode) {
 - (Line*)lineAt:(NSInteger)index;
 - (void)handleTabPress;
 
-- (void)moveSelectedLinesDown;
-
 - (NSInteger)getPageNumber:(NSInteger)location;
 - (NSInteger)numberOfPages;