+++ Beat/User Interface/Editor View/BeatTextView.m
@@ -356,11 +356,7 @@ static NSTouchBarItemIdentifier ColorPickerItemIdentifier = @"com.TouchBarCatalo
 			return; // Skip default behavior
 		case 125:
 			// Down
-			if (theEvent.modifierFlags & NSEventModifierFlagOption) {
-				NSLog(@"Move line down");
-				[self.editorDelegate moveSelectedLinesDown];
-			}
-			else if (self.autocompletePopover.isShown) {
+			if (self.autocompletePopover.isShown) {
 				if (row + 1 >= self.autocompleteTableView.numberOfRows) row = -1;
 				[self.autocompleteTableView selectRowIndexes:[NSIndexSet indexSetWithIndex:row+1] byExtendingSelection:NO];
 				[self.autocompleteTableView scrollRowToVisible:self.autocompleteTableView.selectedRow];